requirements.txt
mysql-connector-python==8.4.0
websockets==12.0
